[FIX] prettier for v13 broken in mirror
https://github.com/pre-commit/mirrors-prettier/issues/6 affects our repos, as it can be seen on https://github.com/OCA/social/runs/1770760744?check_suite_focus=true#step:4:110 for https://github.com/OCA/social/pull/664 and it produces random results when running `pre-commit` in different environments.

I apply here the workaround explained in that issue.

At the same time, I double-checked that we can run all prettier jobs in a single hook, so I take this chance to apply that logic, and to limit by file extensions, to make it work just like the v14 branch with #48, more predictable and efficent.